+++ src/collections.ts
@@ -1,5 +1,6 @@
 import type { Annotation, Decoder } from '~/core';
 import { annotate, annotateObject, formatShort, merge } from '~/core';
+import { quote } from '~/lib/text';
 
 import { array } from './arrays';
 import { pojo } from './objects';
@@ -25,14 +26,12 @@ export function record<K extends string, V>(
     let rv = {} as Record<K, V>;
     const errors: Map<string, Annotation> = new Map();
 
-    for (const [key, value] of Object.entries(input)) {
+    for (const key of Object.keys(input)) {
+      const value = input[key];
       const keyResult = keyDecoder?.decode(key);
       if (keyResult?.ok === false) {
         return err(
-          annotate(
-            input,
-            `Invalid key ${JSON.stringify(key)}: ${formatShort(keyResult.error)}`,
-          ),
+          annotate(input, `Invalid key ${quote(key)}: ${formatShort(keyResult.error)}`),
         );
       }