+++ src/objects.ts
@@ -3,6 +3,7 @@
 import type { Annotation, Decoder, DecodeResult, DecoderType } from '~/core';
 import { annotateObject, define, merge, updateText } from '~/core';
 import { difference } from '~/lib/set-methods';
+import { quote } from '~/lib/text';
 import { isPojo } from '~/lib/utils';
 
 type RequiredKeys<T extends object> = {
@@ -124,9 +125,7 @@ export function object<Ds extends Record<string, Decoder<unknown>>>(
       }
 
       if (missingKeys.size > 0) {
-        const errMsg = Array.from(missingKeys)
-          .map((key) => `"${key}"`)
-          .join(', ');
+        const errMsg = Array.from(missingKeys).map(quote).join(', ');
         const pluralized = missingKeys.size > 1 ? 'keys' : 'key';
         objAnn = updateText(objAnn, `Missing ${pluralized}: ${errMsg}`);
       }
@@ -157,13 +156,12 @@ export function exact<Ds extends Record<string, Decoder<unknown>>>(
     const actualKeys = new Set(Object.keys(plainObj));
     const extraKeys = difference(actualKeys, allowedKeys);
     return extraKeys.size > 0
-      ? `Unexpected extra keys: ${Array.from(extraKeys).join(', ')}`
-      : // Don't reject
-        null;
+      ? `Unexpected extra keys: ${Array.from(extraKeys).map(quote).join(', ')}`
+      : null;
   });
 
   // Defer to the "object" decoder for doing the real decoding work
-  return checked.then(object(decoders).decode);
+  return checked.pipe(object(decoders));
 }
 
 /**
@@ -177,9 +175,9 @@ export function inexact<Ds extends Record<string, Decoder<unknown>>>(
 export function inexact<Ds extends Record<string, Decoder<unknown>>>(
   decoders: Ds,
 ): Decoder<ObjectDecoderType<Ds> & Record<string, unknown>> {
-  return pojo.then((plainObj) => {
+  return pojo.pipe((plainObj) => {
     const allkeys = new Set(Object.keys(plainObj));
-    const decoder = object(decoders).transform((safepart) => {
+    return object(decoders).transform((safepart) => {
       const safekeys = new Set(Object.keys(decoders));
 
       // To account for hard-coded keys that aren't part of the input
@@ -200,6 +198,5 @@ export function inexact<Ds extends Record<string, Decoder<unknown>>>(
       }
       return rv;
     });
-    return decoder.decode(plainObj);
   });
 }