+++ test-d/inference.test-d.ts
@@ -10,6 +10,7 @@ import {
   date,
   datelike,
   decimal,
+  define,
   dict,
   either,
   email,
@@ -30,6 +31,8 @@ import {
   mapping,
   maybe,
   mixed,
+  nanoid,
+  identifier,
   never,
   nonEmptyArray,
   nonEmptyString,
@@ -38,7 +41,6 @@ import {
   null_,
   number,
   numeric,
-  numericBoolean,
   object,
   oneOf,
   optional,
@@ -66,9 +68,6 @@ import {
   // Formatters
   formatInline,
   formatShort,
-
-  // Results
-  ok,
 } from '../dist';
 import { expectError, expectType, expectAssignable } from 'tsd';
 
@@ -85,6 +84,15 @@ function foo(
   _r: DecoderType<typeof strings>,
   _s: DecoderType<typeof truthy>,
 ) {}
+
+expectType<123 | 'hi'>(
+  test(
+    define((blob, ok, err) => {
+      expectType<unknown>(blob);
+      return Math.random() < 0.5 ? ok(123) : Math.random() < 0.5 ? ok('hi') : err('fail');
+    }),
+  ),
+);
 
 expectType<(p: string, q: number[], r: string[], s: boolean) => void>(foo);
 
@@ -116,6 +124,12 @@ expectType<string>(test(email));
 expectType<string>(test(regex(/foo/, 'Must be foo')));
 expectType<URL>(test(url));
 expectType<URL>(test(httpsUrl));
+expectType<string>(test(identifier));
+expectType<string>(test(nanoid()));
+expectType<string>(test(nanoid({ size: 7 })));
+expectType<string>(test(nanoid({ min: 8 })));
+expectType<string>(test(nanoid({ max: 16 })));
+expectType<string>(test(nanoid({ min: 8, max: 16 })));
 expectType<string>(test(uuid));
 expectType<string>(test(uuidv1));
 expectType<string>(test(uuidv4));
@@ -160,7 +174,49 @@ string.verify('dummy', formatShort);
 expectType<number | undefined>(number.value('dummy'));
 expectType<string | undefined>(string.value('dummy'));
 
-expectType<number>(test(string.then((value: string) => ok(value.length))));
+expectType<number>(test(string.then((value: string, ok) => ok(value.length))));
+expectType<number>(
+  test(
+    string.then((value: string, ok, err) =>
+      Math.random() < 0.5 ? ok(value.length) : err('Nope'),
+    ),
+  ),
+);
+expectType<number | string>(
+  test(
+    string.then((value: string, ok, err) =>
+      Math.random() < 0.3
+        ? ok(value.length)
+        : Math.random() < 0.5
+          ? ok(value)
+          : err('Nope'),
+    ),
+  ),
+);
+
+// .then()
+expectType<number>(test(string.transform(Number).then(positiveInteger)));
+expectType<number>(test(string.transform(Number).then(positiveInteger.decode)));
+expectType<boolean>(test(string.transform(Number).transform(String).then(truthy)));
+expectType<boolean>(test(string.transform(Number).transform(String).then(truthy.decode)));
+
+// .pipe()
+expectType<number>(test(string.transform(Number).pipe(positiveInteger)));
+expectType<boolean>(test(string.transform(Number).transform(String).pipe(truthy)));
+// .pipe() with branch infers decoder from both branches
+expectType<number | string>(
+  test(string.transform(Number).pipe(Math.random() < 0.5 ? positiveInteger : string)),
+);
+// .pipe() with function with branches infers decoder from both branches
+expectType<number | string>(
+  test(
+    string.transform(Number).pipe(() => (Math.random() < 0.5 ? positiveInteger : string)),
+  ),
+);
+// .pipe() with input function with branches infers decoder from both branches
+expectType<number | string>(
+  test(string.transform(Number).pipe((x) => (x < 0.5 ? positiveInteger : string))),
+);
 
 expectType<string>(test(string.refine((s) => s.startsWith('x'), 'Must start with x')));
 
@@ -210,7 +266,6 @@ expectType<unknown[]>(test(poja));
 
 expectType<boolean>(test(boolean));
 expectType<boolean>(test(truthy));
-expectType<boolean>(test(numericBoolean));
 
 expectType<string | undefined>(test(optional(string)));
 expectType<string | undefined>(test(optional(optional(string))));